Split repo into scripts/ + template/ so rename.sh stops mutating in place
rename.sh used to sed every file in the whole repo (including .git,
.github, scripts/) with BSD-only `sed -i ''` syntax, which is broken on
Linux and destructive since it edited the one canonical copy in place.

Template content now lives under template/; rename.sh copies template/
into a new project directory and only renames/edits the copy, leaving
the template untouched and re-runnable. Also fixes the sed portability
bug and a set -e/pipefail interaction that silently aborted the script
partway through. `inv rename` at the repo root still works via a thin
wrapper tasks.py.
