+++ requirements.txt
@@ -1,2 +1,1 @@
 invoke==1.4.1
-requests==2.32.3
