+++ template/VERSION.txt
