+++ template/demos/processor-offline/CMakeLists.txt
