+++ template/demos/processor-realtime/CMakeLists.txt
