+++ template/demos/sink-offline/CMakeLists.txt
