+++ template/demos/source-realtime/CMakeLists.txt
