+++ template/include/ExampleDSPExtension.hpp
