+++ template/requirements.txt
@@ -1,0 +1,2 @@
+invoke==1.4.1
+requests==2.32.3
