+++ template/src/all/CMakeLists.txt
