+++ template/src/all/ExampleDSPExtension.cpp
