+++ template/src/all/ExampleDSPNodeFactory.cpp
