+++ template/src/all/ExampleDSPNodeFactory.hpp
