+++ template/src/all/Nodes/ExampleProcessor/ExampleProcessorNode.cpp
