+++ template/src/all/Nodes/ExampleProcessor/ExampleProcessorNode.hpp
