+++ template/src/all/Nodes/ExampleSink/ExampleSinkNode.cpp
