+++ template/src/all/Nodes/ExampleSink/ExampleSinkNode.hpp
