+++ template/src/all/Nodes/ExampleSource/ExampleSourceNode.cpp
