+++ template/src/all/Nodes/ExampleSource/ExampleSourceNode.hpp
